+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- The MakeServicesPublicAsALastResortCompilerPass that allows to make some of the
+  services public again, after they have been made private in v5.0.0. This is supposed
+  to be a last resort solution (see the compiler pass' docblock).
 
 ## [5.0.0] - 2020-03-30
 ### Changed